Bug fix: empty topic string passed to (un)subscribed signals
Due to changes in commit 83ad60daee086b0fdf5012a10060a195ea062338
the subscribed and unsubscribed signals in the Client class
would be emitted with an empty topic string.

This is due to the fact that the (UN)SUBACK message does not
contain the topic.

This commit binds the topic to the message id when the (UN)SUBSCRIBE
message is sent, and recovers the topic again when the acknowledgement
is received.