Added CMake project configuration.
This adds the ability to use this project via CMake.  The project has the same
options with the same defaults as the Qt project.  It also ups the minimum Qt
version from 5.3 to 5.7 if the web sockets option is enabled.

The CMake project also adds the ability to generate installable packages in the
customary manner - one with only the shared library, and one with development
files, including a generated CMake helper to make it easier to use in other
CMake projects.  Simply run `cmake -G RPM` (or "DEB") after building with CMake.

Caveat:
This version of the CMake project only covers "src" - and _not_ "examples" or "tests".

Testing status:
- RPM tested on OpenSUSE Leap 15.1
- DEB tested on Ubuntu Core 19.04
- NSIS (Windows/setup.exe) not tested, but might work anyway.