Errors found by the PVS studio static analyzer.
Errors found by the PVS studio static analyzer:
qmqtt_client_p.cpp	52	warn	V730 It is possible that not all members of a class are initialized inside the constructor. Consider inspecting: _ignoreSelfSigned.
qmqtt_network.cpp	140	warn	V773 The '_socket' pointer was not released in destructor. A memory leak is possible.
qmqtt_network.cpp	140	warn	V773 The '_autoReconnectTimer' pointer was not released in destructor. A memory leak is possible.